# schemas/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "CreateRunRequest": {
        "properties": {
          "agent_id": {
            "format": "uuid",
            "type": "string"
          },
          "input": {
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "input"
        ],
        "type": "object"
      },
...
      "Run": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "waiting",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cognitive runtime foundation — memory, retrieval, routing, execution, and diagnostics",
    "title": "Vestrace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs": {
      "get": {
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-principal-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of runs"
          }
        },
        "summary": "List runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-principal-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Run created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Create a run",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}